+++ src/DbLive/Project/Exceptions/DuplicateMigrationItemException.cs
@@ -2,9 +2,9 @@
 
 
 [ExcludeFromCodeCoverage]
-public class DuplicateMigrationItemException(int migrationVersion, MigrationItemType migrationItemType)
+public class DuplicateMigrationItemException(long migrationVersion, MigrationItemType migrationItemType)
 	: Exception($"Migration item already defined, check sql migrations in the project. MigrationVersion='{migrationVersion}', MigrationItemType={migrationItemType}.")
 {
-	internal int MigrationVersion { get; private set; } = migrationVersion;
+	internal long MigrationVersion { get; private set; } = migrationVersion;
 	internal MigrationItemType MigrationItemType { get; private set; } = migrationItemType;
 }
